+++ protected/controllers/IngredientStatesController.php
@@ -63,8 +63,14 @@ class IngredientStatesController extends Controller
 		if(isset($_POST['IngredientStates']))
 		{
 			$model->attributes=$_POST['IngredientStates'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->IST_ID));
+			
+			if(Yii::app()->user->demo){
+				$this->errorText = sprintf($this->trans->DEMO_USER_CANNOT_CHANGE_DATA, $this->createUrl("profiles/register"));
+			} else {
+				if($model->save()){
+					$this->redirect(array('view','id'=>$model->IST_ID));
+				}
+			}
 		}
 
 		$this->checkRenderAjax('create',array(
@@ -87,8 +93,14 @@ class IngredientStatesController extends Controller
 		if(isset($_POST['IngredientStates']))
 		{
 			$model->attributes=$_POST['IngredientStates'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->IST_ID));
+			
+			if(Yii::app()->user->demo){
+				$this->errorText = sprintf($this->trans->DEMO_USER_CANNOT_CHANGE_DATA, $this->createUrl("profiles/register"));
+			} else {
+				if($model->save()){
+					$this->redirect(array('view','id'=>$model->IST_ID));
+				}
+			}
 		}
 
 		$this->checkRenderAjax('update',array(
@@ -106,7 +118,11 @@ class IngredientStatesController extends Controller
 		if(Yii::app()->request->isPostRequest)
 		{
 			// we only allow deletion via POST request
-			$this->loadModel($id)->delete();
+			if(Yii::app()->user->demo){
+				$this->errorText = sprintf($this->trans->DEMO_USER_CANNOT_CHANGE_DATA, $this->createUrl("profiles/register"));
+			} else {
+				$this->loadModel($id)->delete();
+			}
 
 			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 			if(!isset($_GET['ajax']))
